refactor: seed simulator tool
