+++ app/templates/base.html
@@ -29,7 +29,7 @@
             <a href="{{ url_for('tools.meal_finder') }}">Meal Finder</a>
             <a href="{{ url_for('tools.tea_helper') }}">Tea Helper</a>
             <a href="{{ url_for('tools.item_helper') }}">Item Helper</a>
-            <a href="{{ url_for('seed_simulator') }}">Seed Simulator</a>
+            <a href="{{ url_for('tools.seed_simulator') }}">Seed Simulator</a>
             <a href="{{ url_for('lecture_assistant') }}">Lecture Assistant</a>
         </div>
         <hr>
