+++ app/tools/routes.py
@@ -1,12 +1,14 @@
+from math import floor
+
 import sqlalchemy as sa
 from app import db
-from flask import jsonify, render_template, request
+from flask import jsonify, redirect, render_template, request
 from sqlalchemy import func
 
-from app.models import liked_meals, liked_gifts, Character, Gift, LostItem, Menu, TeaTopic
-from app.tools.forms import CharacterForm, DualCharacterForm, ItemForm
+from app.models import liked_meals, liked_gifts, Character, Gift, LostItem, Menu, Seed, TeaTopic
+from app.tools.forms import CharacterForm, DualCharacterForm, ItemForm, SeedForm
 from app.tools import bp
-from app.utility import get_choices
+from app.tools.utility import get_choices, get_yield_ratios
 
 
 @bp.route('/meal-finder', methods=['GET'])
@@ -106,3 +108,48 @@ def get_item_data():
     return jsonify({'character': character_data,
                     'gifts': gift_data})
 
+@bp.route('/seed-simulator', methods=['GET', 'POST'])
+def seed_simulator():
+    form = SeedForm()
+    seed_choices = [(-1, '')]
+    seed_choices.extend(get_choices(Seed, False))
+    seed_choices = [(sid, name.removesuffix(" Seeds")) for sid, name in seed_choices]
+    form.seed1.choices = seed_choices
+    form.seed2.choices = seed_choices
+    form.seed3.choices = seed_choices
+    form.seed4.choices = seed_choices
+    form.seed5.choices = seed_choices
+
+    if form.validate_on_submit():
+        return redirect('')
+
+    return render_template('tools/seed_simulator.html', title='Seed Sim',
+                           page_name='Seed Simulator', form=form)
+
+@bp.route('/get_seed_data', methods=['POST'])
+def get_seed_data():
+    sids = [int(request.json[f'seed{i}_selected_option']) for i in range(1, 6)]
+    unique_seeds = db.session.query(Seed).filter(Seed.id.in_(sids)).all()
+    seed_map = {seed.id: seed for seed in unique_seeds}
+    seeds = [seed_map[sid] for sid in sids if sid != -1]
+    cultivation = int(request.json['cultivation_selected_option'])
+
+    rank, grade = 0, 0
+    for seed in seeds:
+        rank += seed.rank
+        grade += seed.grade
+
+    if not seeds:
+        score = 0
+    else:
+        x = (12 - (rank % 12)) * 5
+        y = floor((grade / 5) * 4)
+        z = (cultivation + 4) * 2
+        score = x + y + z
+
+    yld, ratio, coefficient = get_yield_ratios(score)
+
+    return jsonify({'score': str(score),
+                    'yield': str(yld),
+                    'ratio': ratio,
+                    'coefficient': str(coefficient)})
